# docs/swagger.yaml
basePath: /v1
definitions:
  delivery.AuthorizeRequest:
    properties:
      action:
        description: '要執行的操作 (例如: read, write, delete)'
        type: string
      resource:
        description: 要訪問的資源
        type: string
    required:
    - action
    - resource
    type: object
  delivery.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  domain.User:
    properties:
      created_at:
        type: string
      id:
        type: integer
      jwt:
        type: string
      password:
        type: string
      roles:
        items:
          type: string
        type: array
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:5002
info:
  contact: {}
  description: RBAC 權限管理服務 API 文檔
  title: RBAC Service API
  version: "1.0"
paths:
  /auth/authorize:
    post:
      consumes:
      - application/json
      description: 驗證用戶是否有權限訪問特定資源
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 授權請求參數
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/delivery.AuthorizeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 驗證成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 無效的請求參數
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授權訪問
          schema:
            additionalProperties: true
            type: object
        "403":
          description: 權限不足
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 驗證權限
      tags:
      - Auth
  /auth/batchRevoke:
    post:
      consumes:
      - application/json
      description: 批量撤銷多個訪問令牌
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 批量撤銷成功
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 無效的令牌
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 批量撤銷訪問令牌
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: 處理用戶登錄並返回訪問令牌
      parameters:
      - description: 登錄請求參數
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/delivery.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登錄成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 無效的輸入或登錄失敗
          schema:
            additionalProperties: true
            type: object
      summary: 用戶登錄
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: 處理用戶登出
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 登出成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 無效的輸入或登出失敗
          schema:
            additionalProperties: true
            type: object
      summary: 用戶登出
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: 使用刷新令牌獲取新的訪問令牌
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 令牌刷新成功
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 無效的刷新令牌
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 刷新訪問令牌
      tags:
      - Auth
  /auth/revoke:
    post:
      consumes:
      - application/json
      description: 撤銷指定的訪問令牌
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 令牌撤銷成功
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 無效的令牌
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 撤銷訪問令牌
      tags:
      - Auth
  /users/{id}:
    get:
      description: 根據用戶ID獲取用戶詳細信息
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 用戶ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功獲取用戶信息
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: 無效的用戶ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: 用戶未找到
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服務器內部錯誤
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 獲取用戶詳情
      tags:
      - Users
  /users/create:
    post:
      consumes:
      - application/json
      description: 使用提供的用戶名和密碼創建新用戶
      parameters:
      - description: 用戶創建信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/delivery.LoginRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 用戶創建成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 參數驗證失敗
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: 用戶名已存在
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: 服務器內部錯誤
          schema:
            additionalProperties:
              type: string
            type: object
      summary: 創建新用戶
      tags:
      - Users
swagger: "2.0"